+++ app/com/lynxanalytics/biggraph/controllers/BigGraphController.scala
@@ -122,7 +122,8 @@ case class FEAttribute(
   canBucket: Boolean,
   canFilter: Boolean,
   isNumeric: Boolean,
-  isInternal: Boolean)
+  isInternal: Boolean,
+  computeProgress: Double)
 
 case class FEProjectListElement(
   name: String,
@@ -159,7 +160,12 @@ case class ProjectListRequest(path: String)
 case class ProjectSearchRequest(
   basePath: String, // We only search for projects/directories contained (recursively) in this.
   query: String)
-case class ProjectList(path: String, directories: List[String], projects: List[FEProjectListElement])
+case class ProjectList(
+  path: String,
+  readACL: String,
+  writeACL: String,
+  directories: List[String],
+  projects: List[FEProjectListElement])
 case class OperationCategory(
     title: String, icon: String, color: String, ops: List[FEOperationMeta]) {
   def containsOperation(op: Operation): Boolean = ops.find(_.id == op.id).nonEmpty
@@ -187,7 +193,7 @@ case class ForkDirectoryRequest(from: String, to: String)
 case class RenameDirectoryRequest(from: String, to: String)
 case class UndoProjectRequest(project: String)
 case class RedoProjectRequest(project: String)
-case class ProjectSettingsRequest(project: String, readACL: String, writeACL: String)
+case class ACLSettingsRequest(project: String, readACL: String, writeACL: String)
 
 case class HistoryRequest(project: String)
 case class AlternateHistory(
@@ -242,6 +248,7 @@ object BigGraphController {
 }
 class BigGraphController(val env: BigGraphEnvironment) {
   implicit val metaManager = env.metaGraphManager
+  implicit val entityProgressManager: EntityProgressManager = env.dataManager
 
   val ops = new Operations(env)
 
@@ -253,6 +260,8 @@ class BigGraphController(val env: BigGraphEnvironment) {
     val visible = projects.filter(_.readAllowedFrom(user))
     ProjectList(
       request.path,
+      dir.readACL,
+      dir.writeACL,
       visibleDirs.map(_.path.toString).toList,
       visible.map(_.toListElementFE).toList)
   }
@@ -279,6 +288,8 @@ class BigGraphController(val env: BigGraphEnvironment) {
 
     ProjectList(
       request.basePath,
+      dir.readACL,
+      dir.writeACL,
       dirs.map(_.path.toString).toList,
       projects.map(_.toListElementFE).toList)
   }
@@ -316,6 +327,7 @@ class BigGraphController(val env: BigGraphEnvironment) {
   def createProject(user: serving.User, request: CreateProjectRequest): Unit = metaManager.synchronized {
     assertNameNotExists(request.name)
     val p = ProjectFrame.fromName(request.name)
+    p.assertParentWriteAllowedFrom(user)
     setupACL(request.privacy, user, p)
     p.initialize
     if (request.notes != "") {
@@ -325,22 +337,25 @@ class BigGraphController(val env: BigGraphEnvironment) {
 
   def createDirectory(user: serving.User, request: CreateDirectoryRequest): Unit = metaManager.synchronized {
     assertNameNotExists(request.name)
-    val d = ProjectDirectory.fromName(request.name)
-    setupACL(request.privacy, user, d)
+    val p = ProjectDirectory.fromName(request.name)
+    p.assertParentWriteAllowedFrom(user)
+    setupACL(request.privacy, user, p)
   }
 
   def discardDirectory(user: serving.User, request: DiscardDirectoryRequest): Unit = metaManager.synchronized {
     val p = ProjectDirectory.fromName(request.name)
-    p.assertWriteAllowedFrom(user)
+    p.assertParentWriteAllowedFrom(user)
     p.remove()
   }
 
   def renameDirectory(user: serving.User, request: RenameDirectoryRequest): Unit = metaManager.synchronized {
-    val p = ProjectDirectory.fromName(request.from)
-    p.assertWriteAllowedFrom(user)
     assertNameNotExists(request.to)
-    p.copy(ProjectDirectory.fromName(request.to))
-    p.remove()
+    val pFrom = ProjectDirectory.fromName(request.from)
+    pFrom.assertParentWriteAllowedFrom(user)
+    val pTo = ProjectDirectory.fromName(request.to)
+    pTo.assertParentWriteAllowedFrom(user)
+    pFrom.copy(pTo)
+    pFrom.remove()
   }
 
   def discardAll(user: serving.User, request: serving.Empty): Unit = metaManager.synchronized {
@@ -372,13 +387,14 @@ class BigGraphController(val env: BigGraphEnvironment) {
   }
 
   def forkDirectory(user: serving.User, request: ForkDirectoryRequest): Unit = metaManager.synchronized {
-    val p1 = ProjectDirectory.fromName(request.from)
-    p1.assertReadAllowedFrom(user)
+    val pFrom = ProjectDirectory.fromName(request.from)
+    pFrom.assertReadAllowedFrom(user)
     assertNameNotExists(request.to)
-    val p2 = ProjectDirectory.fromName(request.to)
-    p1.copy(p2)
-    if (!p2.writeAllowedFrom(user)) {
-      p2.writeACL += "," + user.email
+    val pTo = ProjectDirectory.fromName(request.to)
+    pTo.assertParentWriteAllowedFrom(user)
+    pFrom.copy(pTo)
+    if (!pTo.writeAllowedFrom(user)) {
+      pTo.writeACL += "," + user.email
     }
   }
 
@@ -394,8 +410,8 @@ class BigGraphController(val env: BigGraphEnvironment) {
     p.redo()
   }
 
-  def changeProjectSettings(user: serving.User, request: ProjectSettingsRequest): Unit = metaManager.synchronized {
-    val p = ProjectFrame.fromName(request.project)
+  def changeACLSettings(user: serving.User, request: ACLSettingsRequest): Unit = metaManager.synchronized {
+    val p = ProjectDirectory.fromName(request.project)
     p.assertWriteAllowedFrom(user)
     // To avoid accidents, a user cannot remove themselves from the write ACL.
     assert(user.isAdmin || p.aclContains(request.writeACL, user),