+++ web/app/scripts/project-selector.js
@@ -40,7 +40,7 @@ angular.module('biggraph').directive('projectSelector', function(util, hotkeys, 
       });
 
       scope.util = util;
-      function refresh() {
+      scope.reload = function() {
         abandonScalars();
         if (!scope.searchQuery) {
           scope.data = util.nocache('/ajax/projectList', { path: scope.path });
@@ -53,13 +53,16 @@ angular.module('biggraph').directive('projectSelector', function(util, hotkeys, 
             });
         }
         window.localStorage.setItem('last_selector_path', scope.path);
-      }
+      };
 
-      scope.$watch('path', refresh);
-      scope.$watch('searchQuery', refresh);
+      scope.$watch('path', scope.reload);
+      scope.$watch('searchQuery', scope.reload);
       function getScalar(title, scalar) {
+        if (scalar.computeProgress !== 1.0) {
+          return NOT_CALCULATED;
+        }
         var res = util.get('/ajax/scalarValue', {
-          scalarId: scalar.id, calculate: false
+          scalarId: scalar.id
         });
         res.details = { project: title, scalar: scalar };
         return res;
@@ -67,6 +70,13 @@ angular.module('biggraph').directive('projectSelector', function(util, hotkeys, 
 
       // Fake scalar for projects with no vertices/edges.
       var NO = { string: 'no', $abandon: function() {} };
+
+      // Placeholder when the scalar has not been calculated yet.
+      var NOT_CALCULATED = {
+        $statusCode: 404,
+        $error: 'Not calculated yet',
+        $abandon: function() {},
+      };
 
       scope.$watch('data.$resolved', function(resolved) {
         if (!resolved || scope.data.$error) { return; }
@@ -82,7 +92,7 @@ angular.module('biggraph').directive('projectSelector', function(util, hotkeys, 
       });
 
       function abandonScalars() {
-        if (scope.data && scope.data.$resolved) {
+        if (scope.data && scope.data.$resolved && !scope.data.$error) {
           for (var i = 0; i < scope.data.projects.length; ++i) {
             var p = scope.data.projects[i];
             scope.vertexCounts[p.name].$abandon();
@@ -182,17 +192,20 @@ angular.module('biggraph').directive('projectSelector', function(util, hotkeys, 
         rename: function(kind, oldName, newName) {
           if (oldName === newName) { return; }
           util.post('/ajax/renameDirectory',
-              { from: oldName, to: newName }).then(refresh);
+              { from: oldName, to: newName }).then(scope.reload);
         },
         duplicate: function(kind, p) {
           util.post('/ajax/forkDirectory',
-              { from: p, to: scope.dirName(p) + 'Copy of ' + scope.baseName(p) }).then(refresh);
+              { 
+                from: p,
+                to: scope.dirName(p) + 'Copy of ' + scope.baseName(p)
+              }).then(scope.reload);
         },
         discard: function(kind, p) {
           var message = 'Permanently delete ' + kind + ' ' + p + '?';
           message += ' (If it is a shared ' + kind + ', it will be deleted for everyone.)';
           if (window.confirm(message)) {
-            util.post('/ajax/discardDirectory', { name: p }).then(refresh);
+            util.post('/ajax/discardDirectory', { name: p }).then(scope.reload);
           }
         },
       };